utils/image_utils.py
import cv2
import numpy as np

def load_image_from_bytes(file_bytes):
    arr = np.frombuffer(file_bytes, np.uint8)
    return cv2.imdecode(arr, cv2.IMREAD_COLOR)

def preprocess_for_ocr(img):
    """Améliore le contraste pour l'écriture manuscrite."""
    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
    # Débruitage
    denoised = cv2.fastNlMeansDenoising(gray, h=10)
    # Binarisation adaptative (meilleure pour manuscrit)
    binary = cv2.adaptiveThreshold(
        denoised, 255,
        cv2.ADAPTIVE_THRESH_GAUSSIAN_C,
        cv2.THRESH_BINARY, 11, 2
    )
    return binary

def crop_zone(img, x_pct, y_pct, w_pct, h_pct):
    """Découpe une zone par pourcentage de la taille totale."""
    h, w = img.shape[:2]
    x1 = int(w * x_pct);  y1 = int(h * y_pct)
    x2 = int(w * (x_pct + w_pct))
    y2 = int(h * (y_pct + h_pct))
    return img[y1:y2, x1:x2]