requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-marshmallow
marshmallow-sqlalchemy
opencv-python
numpy
pillow